Split indexing util funcs to a separate file
